+++ src/main.py
@@ -1,12 +1,12 @@
 from itertools import chain
 
-from model.any2vec import Any2Vec
-from model.anys.tropes import get_tropes_from_file, get_unique_tropes, get_training_tropes
-from model.anys.words import get_sentences_from_file, get_corpus_from_sentences, get_words_in_vocabulary, \
+from any2vec.any2vec import Any2Vec
+from any2vec.anys.tropes import get_tropes_from_file, get_unique_tropes, get_training_tropes
+from any2vec.anys.words import get_sentences_from_file, get_corpus_from_sentences, get_words_in_vocabulary, \
     generate_training_data
-from model.data import LearningRate
-from model.model_runner import ModelRunner
-from model.vocabulary import Vocabulary
+from any2vec.data import LearningRate
+from any2vec.model_runner import ModelRunner
+from any2vec.vocabulary import Vocabulary
 
 
 def any2vec_with_words(embedding_size: int, learning_rate: LearningRate) -> [str, Any2Vec]:
