+++ src/tests/model/anys/test_words.py
@@ -1,9 +1,9 @@
 import unittest
 import os
 
-from model.anys.words import get_sentences_from_file, get_corpus_from_sentences, get_words_in_vocabulary, \
+from any2vec.anys.words import get_sentences_from_file, get_corpus_from_sentences, get_words_in_vocabulary, \
     generate_training_data
-from model.training_data import TrainingData
+from any2vec.training_data import TrainingData
 
 
 class WordsTestCase(unittest.TestCase):
